Mat code for the basic LUFS meter
outputs three sets of data, Momentary, Short term and integrated LUFS.
This follows the EBU r128 specification, except the useless stuff like the thing about the decimal points, and being real time as opposed to across a file. This one works for the whole file. 
This version is only stereo L and R. 
I will work to add a UI, different EBU offset variations and other bits in the coming months.